feat: add --auto-commit command line option to RUN
This change adopts the revision of compiler.AutoCommit() to include
a flag for all vs. requied automatic imports. Additionally, the
"run" command is extended to include an --auto-import boolean value
that can override the default profile setting. This would normally
be used for debugging a problem with the auto-import processing.