add correct HTTP status codes for services.
A service program can use the @status directive to set the rest
status to an arbitrary integer value.

Authentication policy is done in the service itself; if it is
missing an authenticated username:password or token, it must
use @status 401 to report it, which also causes a header to be
added so a browser will issue a password prompt and retry.

Logging is enabled in the server by default, you must use the
--no-log command line option to supress the logging.