feat: cache service programs for performance
The REST server will cache compilation units in memory of
services it has compiled to support endpoints. Each time
an endpoint is loaded, if it isn't in the cache it is
compile and saved in the cache.

IF the cache grows too large, the oldest item(s) will be
removed from the cache. Use --cache-size to set the maximum
number of service programs to cache.