Add SERVER START and SERVER STOP
The "start" command will detach a process and start a server. The
output is redirected to the file "ego-server.log" in the current
directory. The file /tmp/ego-server.pid is used to hold the pid of
the last server started.

The "stop" command will kill the process if it already exists, and
will delete the pid file.