Improved log and user db handling
The log and user database file names are converted to absolute
file paths and stored as part of the persistent state. This means
restarts will use the same files regardless of where the restart
command is issued.

Also, set the default permissions for the pid file and the user
database file to be 0600 so only the server user can read or
delete them.