fix: better handling of this for chained fns
r := rest.New().Media("text").Base(url)   wasn't working, because the
"this" veriable was only set on the first call (to "rest"). But that fn
returns an object, and the chained call must re-assert the "this"
as the previous calls' function result for the chain to work.

This fix removes the "predictive" nature of the SetThis operation
in favor of making it smarter when member dereferences occur
in anticipation of a call.