feat: array casting from types
You can use type syntax to specify a cast. For example,
a := []int("test")  will convert the string into an array of
int values, each representing a rune.  Similarly, string(a)
where a is an array of ints returns the string made up of
those runes.

You can also cast an array from one type of array value
to another, such as []float(i) to convert an array of ints
to an array of floats, etc. Int, float, bool, and strings are
supported for this kind of array conversion.

Added an integration test to validate this, and fixed
a unit test for members() and reflect() that were not
expecting Ego arrays properly.