feat: more complete user package support
This allows you to specify a type in a package that can be referenced
outside the package, using the package name. Additionally,
receiver functions in the package are available to the type(s) defined
in the package.

Adds the ability to create instances of package types, including with
initializer values.

Removed deprecated code that allowed symbols to be stored in bytecode
and simplified how package symbols are managed.