perf: more efficient symbol table memory mgmt
Instead of having a single fixed array of symbol values for a given
scope, the symbol table as an array of bins, each of which has
a fixed array. The bins can be extended indefinitely.

A larger value means fewer bins but may waste memory; a
smaller value means more bins and slice appends.