feat: allow credential store via postgres
If the user database starts with "postgres://" it is a URL
for the Postgres client. Otherwise, the name is used as a
physical filename path.

The database URL must include a database name. If a table
"credentials" does not exist in the table, it is created.

Additional processing is done in the server start command to
only apply filename path expansion when it is not a postgres
database location.