fix: issues with json of native maps and arrays
Native (i.e. datatypes.EgoMap and datatypes.EgoArray) items were not being handled
by the json.Marshal() and json.UnMarshal() functions correctly.

Note that in order to preserve maximum flexibility with REST clients,
the metadata elements of a map or array are lost when converting to
json and back; the result of a map will always be a struct and the
result of an array will always be an anonymous array.