fix: json Unmarhsal  to typed value
With strong(er) typing, json's tendency to unmarshal into a
map in most cases was untenable. An optional second argument
to json.Unmarshal() must be a pointer to an instance of the
desired type to store into. This parrots the Go version of the
same functionality.

Note also that Unmarshal has had it's name changed to make
it more compatible with the Go version.