fix: allow initializer values to be arbitrary expressions
In addition to the "{...}" literal notation, you can also just specify a
value expression as the intiializer value. So you can then mix and
match literal and computed parts of an initialization.

This also restores the ability to specify type expressions in the
value field that contain their own initializers.