chore: simplify compound tokenization
Ego uses the standard Go language scanner for tokenization. However,
the Ego compiler wants each token to be a complete Ego token. So as
the tokens are scanned, they are compared against a dictonary
of tokens that are to be "crushed" into a single token. For example,
"<" followed by "=" becomes a single token, "<=".