feat: add sqlite3 support
You can open a database using sqlite3://<filename> to access a sqlite3
database with the given file. The file is created if it does not exist.
All database operations are supported with sqlite that are supported
with postgres.

Split the integration test into two tests, one for each database type.