feat: service handlers have explicit write
Previously, any output to "Stdout" went to the response buffer for the
service. This made it difficult to differentiate between logging
and debug messages, and actual content. This change adds a
'Response' object as the second handler function parameter. The
handler should use this to WriteStatus(), Write(), or WriteJSON() to
the response buffer. All other output goes to the server log.