fix: require basic blocks for func, for, if
Previously, the body of a for-loop, a function, or the if or else
clauses of an if statement could be either a basic block or a
statement. This makes for unreadable code, and deviates from
the Go standard unnecessarily. This updates the compiler to
require a basic block for the statement body.