feat: better json media handling
A service does not need to differentiate between json and text unless
the information must be specially formatted by the service for
text.  The resp.Write() function will automatically handle the
marshalling of JSON data of the object written if it is a single
item.

This also correctly sets the Content-Type to application/json when json
data is requested.