chore: simplify logging identifiers
Loggers had been identified by name everywhere in the code; this
required passing a string to each logging call that was looked
up in a map. To simplify this and remove the map, the logging
status is stored in an array indexed by a logger class ID which is
an integer.

Added Logger(string) and LoggerName(int) to map names
to logger IDs for UI cases such as the -d command line option.