finish support for native Ego http
This required some addiitonal symbol table caching to accompany the
bytecode cache, to retain type and function definitions properly from
imports. This allows a service to use Ego-native http definitions, and
be cached in the server.

This also adds an INFO logger, and moves much of the fine-grained
detail out of the SERVER logger and into the INFO logger (things like
cache status) so a default server log really just records a line for
each request showing the URL and status.