unify UP message for local vs. remote
The "ego server status" message is now the same for a server that is UP
whether it is examlining the local process via the PID file, or calling
the UP endpoint on the server. The keyword LOCAL is added to the
end of the message when it is the local process being reported upon.