feat: allow host[:port] for logging
You can specify the server address (and optionally port) on the logging
enable or disable operation to direct the command to a specific server.
If not specified, it defaults to the current application server, or
logon server if no application server is defined. Finally, defaults
to localhost:8080 if no other defintiions.