feat: rolling log files
When you start ego and specify --log with a log file name, that name is
modfiied with a timestamp and log messages are written there. Each night
at midnight that the log is still opened, the log is closed and a new
log file (with a new datestamp in the name) is created. This prevents
server restarts, etc. from erasing previous log data.

Added the egolog shell script to make it easy to look at the most recent
log file.