fix: broken system packages
If a system package was extended in REPL mode (such as the Camel
function in strings), it was not updating the system package info. This
correct that; any import can update a system package.

Also fixed issue where asset reads of directories result in a more
secure error message that doesn't expose internal file system
path names.