feat: sandboxing file paths
If the ego.sandbox.path profile setting is enabled, it defines the path
within which all file system operations are limited for Ego functions,
such as io.ReadDir() or io.WriteFile().

For server mode, this can also be set using --sandbox-path <path>
on the command invocation of the server. Sandboxing of the path is
not supported in test mode.