feat: add support for localization in Ego programs
Use the new @localization directive to define named
property strings for messages for each supported
language (EN, ES, FR, etc).

Use the i18n.T() function to get the proper localized
message for the current language. Operationally,
specify parameters that can be passed into the string
when it is resolved.