fix: allow @main before body of program
The @main directive is used to identify the name of the main
program function. If not specified, it defaults to "main". This
fix allows the @main directive to appear anywhere in the program
body (outside a function definition) and it will take effect when
the end of the source file is read and the main program is called.

The name of the program is stored in a hidden global variable and
used by the main function call operation appended to the end of
the ego program during compilation.