Create default for insecure server
Create ego.server.insecure boolean preference. This defaults
to "true" for a new instance of ego. This is the default state
of whether a server runs insecurely. By default, the server
runs insecurely until this option is changed.

This supports early development. In the future, this should have
the default change to "false" to
assume a secure server (and the associated required certs, etc.)