json support for logon service
Previously, the only response was a text payload with the token. However
if the caller specifies application/json as the accept type, the reply
will be a JSON payload with additional data for issuer, expiration date,
and identity of the token holder.

The `ego logon` command has also been updated accordingly to use the
JSON representation in its rest call, and reports the token expiration
as part of the logon results message.