Thread-safe package lists in compiler
There were circumstances where launching a lot of go routines at
once could result in a map safety issue with the hashmap that
describes the active pacakges for a given thread. This change
protects all package operations with a RWLock.

Note that currently there is a single lock, and all package
operations are serialized. Later, the mutex can be moved into
the package once package-copy operations are cleaned up.