Support insert and update of rows via CLI
To insert a row, specify the table name and then column=value
parameters for the columns to insert. Columns in the table not
named in the command will receive a null value.

ego table insert mydata age=55 name="Susan"

To update rows, specify the table, column value to change, and
an optional filter. If the filter is not specified, the change
is applied to all rows in the table. Only the named column(s)
are modified; other columns in the row remain unchanged.

ego table update mydata age=40 --where 'name="Susan"'