Add ability to execute direct SQL
./ego tables sql ['direct SQL'] [--sql-file text-file-name]

If SQL is not on command line and file not specified, will prompt
console for input until a blank line is entered to form the SQL.
SQL requires admin privileges.

Uses new API,

PUT /tables/@sql

Note use of "@sql" instead of a table name. The body must be a
JSON-encoded string containing the SQL to execute. The result
is either a DBRows or a DBRowCount depending on wheter it is a
SELECT statement or something else.