Make service cache size a preference item
ego.server.cache.size is an integer indicating the initial service
cache size when running as a REST server. If not specified, the
default is zero, with no caching.

Note that currently, there is a problem with referencing package
items from cached services. When this is resolved, the default
cache size should be changed back to 10.