Change in handling of schema and table name
When an unqualified table name is given to the REST API, it is
converted to a fully-qualified name using the user name. All SQL
is then executed with a full schema.name, and all checks for
permissions are done using schema.name.

Similarly, the list operaiton for tables now returns an array
of objects, not strings. Each object contains the schema and
name of the table.