Add CLI interface for creating table
./ego tables create <name> column:type [column:type...]

The types must be the Ego type names of the items to process. You
can optionally put ",nullable" after the type name to indicate
this is a nullable column. note that the column definitions must
be quoted if they contain spaces or other special characters.

Also, fixed issue where privileges table was not being created if
needed, and there was a confusion in the name (permissions versus
privileges) in the builtin DML that has been fixed.