Add additional permissions REST apis
GET    /tables/@permissions         - list all permissions
PUT    /tables/{table}/permissions  - Set permissions to JSON string array
DELETE /tables/{table}/permissions - Remove all permissions

Note that for PUT, the array elements can have a "+" or "-" prefix
to indicate if the privilege is to be added or deleted. It is not
an error to delete a non-existent privilege.