Add pagination to table rows GET
You can now specify ?limit=n to limit the size of the result set,
and ?start=n to specify the starting row in the result set to
return as part of the response payload. You can page through a
result set by adding the limit to the previous start to get the
next start value. A start value is 1-based.