Cleaner REST client error handling
When a REST call returns a non-success HTTP status, and the
response payload contains a field "msg" or "message", then
that text is used as the error message from the Exchange
call. This prevents client code from having to do explicit
error retrieval.

This mode is controlled by the ego.runtime.rest.error setting,
which defaults to true. Set this to false to rever to the old
behavior which requires the client to do the error processing.