Client-side cleanup of status reporting
When an API call is made to support a client function (like listing
availabel usernames), the representation used in the REST call no
longer includes the RestResponse. That type is only used when the
REST call reports a non-success error.

So the addendum to the representations was deleted, cleaning up
the successful response payload definitions.

Additionally, "users list" was converted to use runtime.Exchange
instead of the direct calls to resty interfaces.