Purge old log files
When a new log file is created, the server will check to see if
there are older files that should be purged. The number of logs
retained is set as the --keep-logs parameter, or as a profile
setting in ego.server.retain.logs. The default is three.

The purge operation is done when a server starts, the log rolls
over due to timestamp, or the value is reset by REST API to the
/admin/loggers endpoint.

If you send a DELETE rest call to the loggers endpoint, it will
triggers a purge operation. The keep value can be updated using
the ?keep option on the DELETE.