Better CLI handling of expired logon token.
If the logon token has expired, this pre-emptively checks that
before using the token in a REST call. If the token is expired,
it puts out a clearer message indicating that the user must
log on again before issuing the command.

The expiration date for the last logon is stored in the user
settings file with the key ego.logon.token.expiration. If this
key is missing, then no expiration check is done, and if the
token has expired, the user will only get a "not authorized"
response from the REST call.