Add row counts to tables listing
By default, the table list REST operation includes the number of
rows in the table. A new parameter to the REST call, ?rowcounts=false,
will suppress this if it becomes a performance problem.

Updated the API doc to reflect this. The effect on the CLI is that
the tables list command now returns rows by default.