Support POST for logon with credentials in request body.
If you authentiate using GET, you must use Basic authentiation to
pass the username and password. If you use POST to the logon
endpoint, you can instead place the username and password in the
request body. For SSL/TLS communication, this helps ensure that
the credentials are always fully encrypted when being sent to the
logon service.

Update the API doc to reflect this. Modified the ego CLI logon
command to use the POST form of the operation. Added debugging
messages when the REST debugger is enabled for the logon command.