Add bulk-insert to PUT /tables/name/rows
You can still insert an object by just sending the single object.
But you can also send a rowset object with an array of objects
in it, and the insert will occur for all rows as a single
transaction.

Additionally, the _row_id_ is ALWAYS reset with an insert, even if
one is already present in the result set presented.