Support bulk update in tables /rows API
If the UPDATE operation is a single object, it behaves as before
(where the update is applied to all rows in the filter, or all rows
if there is no filter).

You can do a "bulk update" by sending a rowset. Each item in the
row set is updated using the provided filter.

More importantly, if the update rowset includes a _row_id_ value,
then that value is implicitly used to filter that row's update.
That is, if you send a rowset that includes row IDs the effect
is that you don't need any further filtering as the changes will
only be made to the given rows.

When a "bulk update" is done, it is done atomically as part of a
single transaction.