Add unique constraint to table create options
When you create a table, a new field "unique" with a boolean value
specifies that the column requires that the value be unique. If not
specified, or false, then the column does not have to be unique.

The CLI is updated to allow this specification in the column type
specification by putting a comma after the type and then the word
unique.

Finally, attempting to insert something that violates the uniqueness
requirement will return an http status of 409, "Conflict" and a
message indicating it was a unique key value constraint error.