Display uniqueness constraint in table metadata
Using the /tables/name/ endpoing to get the metadata about a table,
the payload now includes the attribute "unique" which is set to true
if the column must be unique. This value is omitted from the payload
if it is not true.

Similarly, the ego tables show CLI command has been extended to
display the uniqueness attribute. Along the way, the output of this
command was modified to correclty support the "indented" format.