Add media type support
First pass at adding media types, to the /tables endpoint support. This
will validate that the media type in the "Accepts" header is valid for
the endpoint called. Additionally, for endopints that return rowsets,
the media type can select between the regular json rowset or the
abstract row format.

Also fixed a bug in the T class for Ego unit tests for True or False
tests.