Revised representations for REST calls
All pre-defined REST calls now include a "server" object (usually
at the start of the JSON payload). This object contains the following
fields:

   api      The API version number (an integer, starting with 1)
   id       The UUID of the server instance
   name     The host name the server runs on
   session  The server session number (correlator) starting with 1

These replace the previous "apiVersion" at the start of thge payload,
and that information is how in the server.api field.

Updated server and client handlers to use new format. Also fixed a
bug in the testware for the panic opcode.