Better logging of multi-line strings
When a logger needs to dump out a multi-line string (such as an
incoming or outgoing REST payload), it now uses a utility to
edit the lines to be printed so they include the session ID
as a prefix to each line.

This is needed to ensure that a query of the log for a given
session ID will include multi-line messages for that session
correctly.