Fix bugs in database update, delete, connection leak
1. A new config setting, ego.server.database.empty.fiter" determines
if a DELETE or PATCH operaiton to a database is permitted without a
filter.  When no filter is given, the operation would apply to ALL
rows in the table. This flag indicates if this is allowed or not;
the default is "false" which says you cannot do this operation with
an empty filter.

2. Fixed a leak in database connections; the code was not correctly
closing database connections after each open.