Fix bug in db insert and update parsing
When an insert or update occurs, the command line is parsed to
locate 'column=value' expressions. The tokenizer was not correctly
handling values that had internal spaces, but the overall expression
was properly quoted. For example,

    timestamp="1/2/2022"

This was being incorrectly tokenzied such that teh value was only "1"
which is the first token of the value part of the expression. This
chaange causes the parser to consider everything after the "=" to
be the value string.