Consolidate command output for text, json, indented
Created a single output function that can handle text, json, and
indented formats. This refactors a lot of wasteful code through
the commands package for formulating user output.

Note that the most common pattern still is to evaluate text output
as a special case requiring per-command formatting of human-
readable output, but the JSON formats are all handled by this new
common routine.