Add HAS() table filter to the API
You can use HAS() in the ?filter URL parameter to specify a column
that contains a value, but does not necessarily equal that value.
The HAS() operator allowed a list of one or more values, each of
which is tested to see if it is contained in the given column value.
This is a case-sensitive test.