Added UPDATE to supported table transaction tasks
A transaction can now contain an operation of "update" which updates
rows. The "filters" object is used to select which rows are to be
updated. The optional "filters" object is an array of columns that
are to be extracted from the "data" object to update (if not given,
all items presented in the data object are updated in the rows).

Updated the API doc to include information about transactions.