Better solution to implied "RUN" verb
Previously, if the argument list was empty then a "run" was the
assumed verb. However, if you wanted to use the assumed verb but
with options, such as setting -d debug loggers, it didn't work
and would instead show the help.

This fix means that if no verb is given, the SetDefaultAction()
call in main will specify a default action to execute.