Add util.SetLogging(n,f) to enable/disable logging at runtime
Ego code uses this function by passing a valid logger name in n
and a boolean value in f. The logger is turned on or off based
on the boolean value f. The function's result is the previous
setting of the logger, which allows code to temporarily set the
logging state and then set it back to whatever it was before.