Add transaction support to @SQL table operation
When a PUT (or POST) is made to the /tables/@sql endpoint, it
specifies that the payload is SQL. Previouysly, this was a single
statement expressed as a string.

With this change, the payload can contain multiple statements,
expressed as an array of strings. When there are multiple statements,
they are executed as a single transaction in the datbase. If there
is a SELECT statement, it must be the last statement in the transaction.