Delete that affects no rows permitted in transaction
In a DELETE rest call to the tables service, if no rows are deleted
it is assumed the filter was incorrect and it returns an error. In
the "delete" operation within a larger transaction, a delete that has
no effect is permitted and does not signal an error in the transaction.