Allow emptyError property on table transaction tasks.
If a task in a transaction to read, delete, or update data has a
filter that does not result in any rows being read or modified, the
default behavior is to continue to transaction tasks.

If the caller sets the emptyError property in the task to true,
then if the select, delete, or update task does not read or modify
any rows, then the transaction is considered to have failed, and
the transaction REST call returns a status of 404 ("not found").