Add error conditions to table transaction tasks
A task can now contain an array of error conditions, expressed
using filter syntax. After a step succeeds, if any conditions are
specified they are evaluated, and if true, cause the task to
fail. The error condition can include an http status value and
a message text, which if present are returned to the REST caller
of the transaction.

The conditions have full access to any symbols created for the
transactions. Additionally, a symbol "_rows_" includes the count
of rows affected or read so far in the transaction.