Increment/decrement are extensions when used as expression atoms
This makes use of increment/decrement operators (++ and --) be
language extensions when used in an expression atom. IF extensions
are turned off, then these are onlh allowed as standalone
statements, like traditional Go.

Also, fixed a bug in the atom case, where the value that was used
as the atom was the post-increment value instead of the pre-
increment value.