fix handling of sort of non-default sizes
The sort.SortInts() only works with array of type int. This adds
support for sort.SortInt32s, sort.Int64s(), and sort.Bytes()

Similarly, float32 and float64 sort functions added.

Finally, sort.Sort() is a generic sort function that will sort
any type without constraint. If the data sent to it is not of
a homogeneious type, the type of the first item is used to
set the type of the result items.