Allow more relaxed type checks on array initializers
When initialing an array of any integer type, allow the initializer
values to be arbitrary integer types. That is, []int32{1,2} can be
used to initialize a set of 32-bit arrays using the scalar constant
values which are implicitly int types.

The same relaxed type checking is also supported for float32 and
float64 array initializers for floating point array values.