Add support for util.Exec() function
This is a utility funcction for executing a nataive shell command
from within an Ego program. The parameters are either the entire
command as the first argument, or the preferred form of each
command argument being an additional function argument. The result
of the function is a string array containing the stdout of the
command execution, and an error value.

This function will return a "no privilige" error if the function
is not enabled via the ego.runtime.exec configuration setting,
which must be "true" to use the function.