Change "@main" to "@entrypoint" to make it clearer
This directive is automatically inserted by the run command to
start the entrypoint (normally "main", but could be something
else when using the --entrypoint option). This change makes the
debugger display something like:

  Stepped to:
   test.ego   16: @entrypoint main

This makes it more clear that when you start debugging, you are
about to call the main entrypoint.