feature: add interface type checking
If a parameter specifies an interface type in a function declaration,
we validate the parameter passed in against the interface function
list to determine if the underlying type matches the protocols.

Note that currently this only checks that the function names align;
next step is matching parameter and return types for the functions
defined in the interface against the functions in the user type
being validated.