feat: field lists in struct declaration
You can now declare a struct with comma-separated lists of
field names followed by a type, which is applied to all the
fields in the list.  i.e.

   type foo struct {
      height, width int64
   }

which is semantically identical to

   type foo struct {
      height int64,
      width  int64,
   }

With this push, also updated the caching of the interface
conformance test to be independent of the specific type,
to prevent unnecessary per-type synching.