remove constants as separate map in symbol tables
Previously, constants were always stored in a separate map from
symbol values. This meant that checking to see if a value was a
constant required chasing the tree of symbol tables examining
each constants map, before being able to set a value.

This uses the new symbols attribute to determine if a symbol is
readonly or not, and uses that and only that to determine if a
symbol can be written to, making the Set family of operations
much faster.