Support implied start and end of slice references
So, this allows a[:5] and a[n:] constructs, where the start or
end of the slice is implicitly either zero ro the length of the
item.

Additionally corrected a problem where the slice of an array was
not returned to the language as an array type, but rather was
a true native slice which, while possible to operate on, was
not possible to compare correctly.

Added a unit test to demonstrate the slice functionality.