Allow server status with expired token
The default rest client function Exchange will test to see if the
current token expiration has passed before trying to send the
message, so it can indicate that any failure is based on an expired
token, not an incorrect authentication (the default from the server).

This fix changes the user agent string from "admin" to "status" for
cases where the token isn't required, such as the server status
command for a non-local server. When the agent is "status" the
client-side check for an expired token is skipped, since the
server won't be looking at the token anyway.