fix to detect bad subcommand keywords or orphaned parameters
The CLI processor captures unrecognized tokens and stores them as
parameters. However, if we parse something that triggers a branch
to a subcommand grammar, it means those unrecognized tokens aren't
allowed to be parameters (which must be after the last subcommand).
Instead, emit an error message indicating the parameter was really
an unrecongized command verb.