Major cleanup of error messaging
This removes the need to use errors.Nil() to test for a nil value;
it can just be tested outright. This makes tests for Ego and native
errors identical in the logic of the code.

Any call that gets a non-nil return code from a Go runtime function
must wrap it in an errors.EgoError() call to add context and location
info to the error.