Better default server name for ego show server
The previous default if no server was given was the hostname, but
this may not be accurate or complete. So now Ego first tries to
use the configured application server. If not set, it then tries
to use the configured login server. If taht is not set, it will
now default again to the native hostname.