fix bug in clone of structure type
The type definition and name were not being correctly copied to
the cloned instance. This meant that a copy of an object used as
a receiver could not locate the functions for that type (which
are attached to the type definition).

Added a unit test to validate this going forward.