Default to not show UUID for users on server users list command
The UUID is of little use to the end user in most cases, and it
clutters the output. So by default, the UUID of each user is not
shown in text output (it is always still present in the JSON).
You can request the UUID be printed by adding --id (or -i) to the
server users list command.