Allow boolean flags to accept optional value for true or false
A flag (like --id) can be marked as boolean. User input can be:

   --id           Assumes true. If not present, value is false
   --id=true      Value is true
   --id=false     Value is false

Note that all common synonyms for true and false are accepted.
This allows the simple inclusion of the option to enable a
feature normally off by default, but to use the same option
with an explicit value if overridding a default is required.