Rename --debug global option to --log
The origin of the "--debug" flag was to enable debug messages, but
this quickly became a class of message types identified by "loggers".
The "-d" option accepted a list. However, it also conflicts semantically
with the "-d" option in the run command, which invokes the Ego
language debugger.

This fix changes the option name to specify the loggers to "--log",
and adds a global option "--log-file" which identifies the name of
the log file to use fo logging output.