Add logging of stack size and instructions executed to trace log
When running code, if more than zero instructions are executed,
the maximum stack size and the number of instructions are logged
to the TRACE logger.

This also involved a minor cleanup in percolating error conditions
from the execution of a program to the main module; a call to
os.Exit() from an Ego program generates a special error that
is intercepted by the main and converted into a Go os.Exit value.