Support relaxed type checking for integer and floating assignments.
   x := int32(55)
   x = 42

This now works; the value of x is int32(42) as the value to store
is coerced if they are both integer types. Similarly for float32
and float64 types.

This only applies when type checking is relaxed. In strict mode,
the types must match exactly in an assignment, and in dynamic
mode the type of the target variable is changed to match the
assigned value.