Enforce exported symbol checking from package types
Types defined within packages have fields that can only be seen
outside the package if they are capitalized. This adds this
check to the member access function, so reading a field name
that isn't exported results in an error if we're not in the
same package. It also adds it to the StoreIndex bytecode to
prevent writing to an unexported field.