Lighter locking for symbol tables that cannot be shared
Instead of locking every table, allow Ego to decide which tables
are shared. This is typically the root table, the main table
assocated with whatever program is being run, and the root of
the handler table.

Other tables that are created as part of normal scope nesting
are not shared between multiple native go routines/threads and
do not need the (expensive) locking.