Support exec.Command operations on Windows
The executor function requires a Windows-specific modification of
the command line option string array, to inject the cmd.exe shell
image and it's options before the user-specified command.

This is implemented in the new fork package, which includes the
execution functions used by server start commands, plus the
options processing functions that are platform specific.