Better stack printing when ego.runtime.stack.trace=true
When this config option is true (defaults to false), tracing shows
all the elements on the stack on a separate line of output. This is
useful when the stack contents are large or complex, and the
default format (all the stack items on one line) truncates before
the useful data is seen.