Better support for detecting pointer references to readonly values
When a value is stored as a readonly value, it is wrapped in an
Immutable structure. When it is retrieved from storage, it is
unwrapped and used normally. But attempts to write to an existing
value that is an Immutable-wrapped item now results in a runtime
error in the Ego program.