Support multiple pointer references in an lvalue
Formerly, you couldn't have a statement like

   **x = 3

because the compiler was hard-wired to assume an identifier would
follow the first "*". This change causes the lvalue compiler to
recognize "*" as a special case, and compiles the remainder of the
lvalue as an expression that puts the pointer on the stack, followed
by a StoreViaPointer opcode with no operand, which says get the
pointer from the stack instead of a named value.