Reworked how reflection returns data, using reflect.Reflection{}
The reflect.Reflect() function used to return a structure with
arbitrary fields based on the type of information being examined.
It was difficult to programatically determine what was even
available in the structure without knowing in advance.

This change replaces the result with a new type, reflect.Reflection{}
A reflection has accessor methods to get information about the
item. If you request information that does not exist, you get a
nil pointer. The Items() function can be used to get a list of
the non-nil items. A String() method can also be called on the
reflection object to get a text form of the object contents.

Updated the language guide to reflect this change.