Add ability to mark grammar objects as unsupported on a given OS
For example, server commands "start", "stop", and "restart" do not
work on Windows, which has to run Ego as a Windows Service instead
of as a simple detached process. The grammar includes new attributes
for those subcommands indicating they do not run on one or more
platforms, expressed as a string array of possbile runtime.GOOS
values ("windows", "darwin", "linux", etc).

Addtionally, individual options can also be marked as unsupported
on a given platform or set of platforms.

The help output is adjusted to not show these commands in the help
text, and the grammar parser will generate an error if they are
used on the wrong platform.