Encapsulate uses of []interface to data.List
The []interface{} is ambiguous with native uses (such as sql row
sets). This change emcapsulates the use of []interface{} when it is
an internal "arbitrary list" to use the data.List type. This is
internally an []interface{} storage object, but hides the type so
we can't accidentally comingle Lists with arrays of interfaces{}

Also fixed a minor formatting bug in one the examples, and added a
small hack to support printing the "len" function properly.