Allow global --env-config option to read config from env vars
When running in a restricted environment (perhaps without a writable
config area, such as in a container), this option causes Ego to
populate any options using environment variables.

The environment variables are upper-case versions of the settings
names, with the "." replaced with "_". So "ego.runtime.path" is
the environment variable "EGO_RUNTIME_PATH". Any environment
variables that match settings names are applied to the default
configuration on startup.

If the option is not specified, then the configuration scan is not
done, and only variable explicitly identified in the grammar as
being settable by environment variable are evaluated.